@@ -8,20 +8,22 @@ plugins {
     id("com.android.application")
     id("com.google.gms.google-services")
     id("com.google.firebase.crashlytics")
+    id("com.huawei.agconnect")
 }
 
 repositories {
     mavenCentral()
     maven { setUrl("https://jitpack.io") }
+    maven { setUrl("https://developer.huawei.com/repo/") }
 }
 
 android {
-    compileSdk = 33
+    compileSdk = 34
 
     defaultConfig {
-        applicationId = project.property("flavor-demo-webim").toString()
+        applicationId = "ru.webim.demo"
         minSdk = 21
-        targetSdk = 33
+        targetSdk = 34
         versionCode = project.property("versionCode").toString().toInt()
         versionName = project.property("sdkVersionName").toString()
         vectorDrawables {
@@ -63,9 +65,14 @@ dependencies {
     implementation("de.hdodenhof:circleimageview:3.1.0")
     implementation("androidx.appcompat:appcompat:1.1.0")
     implementation("com.google.android.material:material:1.1.0")
+
     implementation("com.google.firebase:firebase-messaging:23.0.0")
     implementation("com.google.firebase:firebase-crashlytics:18.3.7")
     implementation("com.google.firebase:firebase-analytics:21.2.2")
+
+    implementation("com.huawei.hms:push:6.7.0.300")
+    implementation("com.huawei.agconnect:agconnect-core:1.5.2.300")
+
     implementation("com.github.thijsk:TouchImageView:v1.3.1")
 
     implementation("androidx.recyclerview:recyclerview:1.1.0")
