@@ -9,6 +9,7 @@ import android.content.Context;
 import android.content.ContextWrapper;
 import android.content.res.Configuration;
 import android.content.res.Resources;
+import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.Build;
 import android.os.VibrationEffect;
@@ -28,12 +29,18 @@ import android.widget.RelativeLayout;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.cardview.widget.CardView;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.core.view.GestureDetectorCompat;
 
 import com.bumptech.glide.Glide;
+import com.bumptech.glide.load.DataSource;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
+import com.bumptech.glide.load.engine.GlideException;
+import com.bumptech.glide.load.model.GlideUrl;
+import com.bumptech.glide.request.RequestListener;
+import com.bumptech.glide.request.target.Target;
 
 import java.io.File;
 import java.util.Locale;
@@ -217,8 +224,7 @@ public abstract class FileMessageHolder extends MessageHolder {
 
     @NonNull
     protected File createFilesCacheDir() {
-        String directory = "webim";
-        File cacheDir = new File(context.getExternalCacheDir() + "/" + directory);
+        File cacheDir = context.getExternalCacheDir();
         if (!cacheDir.exists()) {
             cacheDir.mkdirs();
         }
@@ -248,8 +254,24 @@ public abstract class FileMessageHolder extends MessageHolder {
 
         String imageUrl = fileInfo.getImageInfo().getThumbUrl();
         Glide.with(context)
-            .load(imageUrl)
-            .diskCacheStrategy(DiskCacheStrategy.AUTOMATIC)
+            .load(new GlideUrlCustomCacheKey(imageUrl, fileInfo.hashCode()))
+            .listener(new RequestListener<Drawable>() {
+                @Override
+                public boolean onLoadFailed(@Nullable GlideException e, Object model, Target<Drawable> target, boolean isFirstResource) {
+                    textView.setText(fileInfo.getFileName());
+                    textView.setVisibility(View.VISIBLE);
+                    messageBody.setVisibility(View.VISIBLE);
+                    layoutAttachedFile.setVisibility(View.GONE);
+                    thumbView.setVisibility(View.GONE);
+                    return false;
+                }
+
+                @Override
+                public boolean onResourceReady(Drawable resource, Object model, Target<Drawable> target, DataSource dataSource, boolean isFirstResource) {
+                    return false;
+                }
+            })
+            .diskCacheStrategy(DiskCacheStrategy.ALL)
             .into(imageView);
         imageView.setOnClickListener(view -> holderActions.onOpenImage(fileInfo.getUrl(), message));
         textView.setText(context.getResources().getString(R.string.reply_message_with_image));
@@ -405,5 +427,19 @@ public abstract class FileMessageHolder extends MessageHolder {
             extension = filename.substring(extensionIndex);
         }
         return messageId + extension;
+    }
+
+    public static class GlideUrlCustomCacheKey extends GlideUrl {
+        private final int uniqueKey;
+
+        public GlideUrlCustomCacheKey(String imageUrl, int uniqueKey) {
+            super(imageUrl);
+            this.uniqueKey = uniqueKey;
+        }
+
+        @Override
+        public String getCacheKey() {
+            return String.valueOf(uniqueKey);
+        }
     }
 }
