@@ -8,19 +8,27 @@ import android.util.Log;
 import android.view.ContextThemeWrapper;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.widget.LinearLayout;
+import android.view.ViewGroup;
+import android.widget.FrameLayout;
+import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.FragmentActivity;
 import androidx.fragment.app.FragmentManager;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import ru.webim.android.sdk.Department;
+import ru.webim.android.sdk.FatalErrorHandler;
+import ru.webim.android.sdk.MessageStream;
 import ru.webim.android.sdk.MessageTracker;
+import ru.webim.android.sdk.Survey;
 import ru.webim.android.sdk.Webim;
+import ru.webim.android.sdk.WebimError;
 import ru.webim.android.sdk.WebimSession;
 import ru.webim.chatview.ChatAdapter;
 import ru.webim.chatview.ChatAdapterImpl;
@@ -28,16 +36,23 @@ import ru.webim.chatview.ChatHolderActionsImpl;
 import ru.webim.chatview.R;
 import ru.webim.chatview.utils.DepartmentsDialog;
 import ru.webim.chatview.utils.FileHelper;
+import ru.webim.chatview.utils.SurveyDialog;
 import ru.webim.chatview.utils.ViewUtils;
 
-public class ChatView extends LinearLayout {
+public class ChatView extends FrameLayout implements FatalErrorHandler {
     private ChatList chatList;
     private ChatEditBar editBar;
+    private ViewGroup chatLayout;
+    private ViewGroup errorLayout;
+    private TextView errorHeader;
+    private TextView errorDescription;
     private WebimSession session;
     private DepartmentsDialog departmentsDialog;
     private String accountName;
     private String locationName;
     private FileHelper fileHelper;
+    private SurveyDialog surveyDialog;
+    private int styleRes;
 
     public ChatView(Context context) {
         this(context, null);
@@ -50,11 +65,10 @@ public class ChatView extends LinearLayout {
     public ChatView(Context context, @Nullable AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
 
-        setOrientation(LinearLayout.VERTICAL);
         TypedArray typedArray = context.obtainStyledAttributes(attrs, R.styleable.ChatView, defStyleAttr, R.style.ChatViewBase);
         accountName = typedArray.getString(R.styleable.ChatView_chv_account_name);
         locationName = typedArray.getString(R.styleable.ChatView_chv_location_name);
-        int styleRes = typedArray.getResourceId(R.styleable.ChatView_chv_chat_style, R.style.ChatViewDefaultStyle);
+        styleRes = typedArray.getResourceId(R.styleable.ChatView_chv_chat_style, R.style.ChatViewDefaultStyle);
         typedArray.recycle();
 
         Context themedContext = new ContextThemeWrapper(context, styleRes);
@@ -63,6 +77,12 @@ public class ChatView extends LinearLayout {
         chatList = rootView.findViewById(R.id.chat_list);
         editBar = rootView.findViewById(R.id.chat_edit_bar);
 
+        chatLayout = rootView.findViewById(R.id.chatLayout);
+        errorLayout = rootView.findViewById(R.id.errorLayout);
+        errorHeader = rootView.findViewById(R.id.errorHeader);
+        errorDescription = rootView.findViewById(R.id.errorDesc);
+
+        initSurveyDialog();
         fileHelper = new FileHelper();
         chatList.setListControllerReady(listController -> editBar.setListController(listController));
     }
@@ -109,6 +129,7 @@ public class ChatView extends LinearLayout {
             chatList.setAdapter(chatAdapter);
         }
         editBar.setStream(session.getStream());
+        setSurveyHandler();
         setDepartmentsHandler();
 
         session.resume();
@@ -125,6 +146,71 @@ public class ChatView extends LinearLayout {
         fileHelper.stopWork();
         session.pause();
         session.getStream().closeChat();
+    }
+
+    @Override
+    public void onError(@NonNull WebimError<FatalErrorType> error) {
+        chatLayout.setVisibility(View.GONE);
+        errorLayout.setVisibility(View.VISIBLE);
+
+        switch (error.getErrorType()) {
+            case ACCOUNT_BLOCKED:
+                showError(R.string.error_account_blocked_header, R.string.error_account_blocked_desc);
+                break;
+            case VISITOR_BANNED:
+                showError(R.string.error_user_banned_header, R.string.error_user_banned_desc);
+                break;
+            case WRONG_PROVIDED_VISITOR_HASH:
+                showError(R.string.error_user_hash_invalid_header, R.string.error_user_hash_invalid_desc);
+                break;
+            default:
+                showError(
+                    R.string.error_unknown_header,
+                    R.string.error_unknown_desc,
+                    error.getErrorString()
+                );
+                break;
+        }
+    }
+
+    private void showError(int errorHeaderId, int errorDescId, String... args) {
+        errorHeader.setText(errorHeaderId);
+        errorDescription.setText(getResources().getString(errorDescId, (Object[]) args));
+    }
+
+    private void setSurveyHandler() {
+        session.getStream().setSurveyListener(new MessageStream.SurveyListener() {
+            @Override
+            public void onSurvey(Survey survey) {
+                if (!surveyDialog.isVisible()) {
+                    try {
+                        FragmentManager fragmentManager = ((FragmentActivity) getContext()).getSupportFragmentManager();
+                        surveyDialog.showNow(fragmentManager, "surveyDialog");
+                    } catch (ClassCastException e) {
+                        Log.e("WebimLog", "Can't get fragment manager");
+                    }
+                }
+            }
+
+            @Override
+            public void onNextQuestion(Survey.Question question) {
+                surveyDialog.setCurrentQuestion(question);
+            }
+
+            @Override
+            public void onSurveyCancelled() {
+                if (surveyDialog.isVisible()) {
+                    surveyDialog.dismiss();
+                }
+                showToast(getContext().getString(R.string.survey_finish_message), Toast.LENGTH_SHORT);
+            }
+        });
+    }
+
+    private void initSurveyDialog() {
+        surveyDialog = new SurveyDialog(styleRes);
+        surveyDialog.setAnswerListener(answer -> session.getStream().sendSurveyAnswer(answer, null));
+        surveyDialog.setCancelListener(() -> session.getStream().closeSurvey(null));
     }
 
     private void setDepartmentsHandler() {
@@ -212,9 +298,14 @@ public class ChatView extends LinearLayout {
 
     private WebimSession createDefaultSession(Webim.SessionBuilder builder) {
         builder.setContext(getContext())
+            .setErrorHandler(this)
             .setAccountName(accountName)
             .setLocation(locationName);
         return builder.build();
+    }
+
+    private void showToast(String messageToast, int lengthToast) {
+        Toast.makeText(getContext(), messageToast, lengthToast).show();
     }
 
     /**
