@@ -6,6 +6,7 @@ import androidx.annotation.Nullable;
 import java.util.List;
 
 import ru.webim.android.sdk.impl.MessageReaction;
+import ru.webim.android.sdk.impl.items.MessageGroup;
 
 /**
  * Abstracts a single message in the message history. A message is an immutable object. It means that for changing
@@ -128,6 +129,12 @@ public interface Message {
     boolean canVisitorChangeReaction();
 
     /**
+     * @return groupId of the message of null if groupId not exists
+     */
+    @Nullable
+    MessageGroup.GroupData getGroupData();
+
+    /**
      * Abstracts unique id of the message. The class was designed only to be compared by {@code equals}.
      * @see Message#getClientSideId()
      */
@@ -204,7 +211,11 @@ public interface Message {
         /**
          * A message had been sent to the server, received by the server and was spreaded among clients.
          */
-        SENT
+        SENT,
+        /**
+         * A message hadn't been sent successfully.
+         */
+        FAILED
     }
 
     /**
@@ -302,6 +313,11 @@ public interface Message {
          * @return if a file is an image, returns information about an image, in other cases returns null
          */
         @Nullable ImageInfo getImageInfo();
+
+        /**
+         * @return local file path
+         */
+        @Nullable String getLocalPath();
     }
 
     /**
@@ -387,7 +403,7 @@ public interface Message {
         /**
          * @return author id of the message that was quoted.
          */
-        @Nullable String getAuthorId();
+        @Nullable String getQuotedMessageId();
 
         /**
          * Shows the state of the quoted message.
